ChartsView UI & Support for average and total steps.

+++ FitnessApp/Charts/Model/DailyStepModel.swift
@@ -10,5 +10,8 @@ import Foundation
 struct DailyStepModel: Identifiable {
     let id = UUID()
     let date: Date
-    let count: Double
+    let count: Int
 }
+
+
+
